linkedin_search/indexer/faiss_store.py
"""
indexer/faiss_store.py
----------------------
Builds, persists, and manages the FAISS IVFFlat index.

Features:
  - IVFFlat index (fast approximate nearest-neighbour at 124K+ scale)
  - Metadata sidecar (JSONL) keyed by FAISS internal ID
  - TTL-based eviction: rebuild index excluding vectors older than N days
  - GCS upload/download hooks for production persistence
"""

import json
import logging
import time
from datetime import datetime, timezone, timedelta
from pathlib import Path
from typing import Any

import faiss
import numpy as np

from config.settings import settings
from scraper.linkedin_scraper import LinkedInPost

logger = logging.getLogger(__name__)


# ── Metadata record ───────────────────────────────────────────────────────────

def _now_iso() -> str:
    return datetime.now(timezone.utc).isoformat()


def _post_to_meta(post: LinkedInPost, faiss_id: int) -> dict:
    return {
        "faiss_id": faiss_id,
        "post_id": post.post_id,
        "url": post.url,
        "author": post.author,
        "author_title": post.author_title,
        "likes": post.likes,
        "posted_at": post.posted_at,
        "indexed_at": _now_iso(),
        "query": post.query,
        "source": post.source,
        "text": post.text,            # stored for retrieval (not re-embedded)
        "roles": getattr(post, "_roles", []),
        "skills": getattr(post, "_skills", []),
        "is_hiring": getattr(post, "_is_hiring", False),
    }


# ── FAISS store ───────────────────────────────────────────────────────────────

class FAISSStore:
    """
    Manages a FAISS IVFFlat index backed by a JSONL metadata sidecar.

    Index layout:
        data/faiss_index/          ← FAISS_INDEX_PATH
            index.bin              ← the FAISS index binary
        data/metadata.jsonl        ← FAISS_METADATA_PATH
    """

    def __init__(
        self,
        embed_dim: int = 384,           # bge-small-en-v1.5 dim
        index_path: Path | None = None,
        meta_path: Path | None = None,
    ):
        self.embed_dim = embed_dim
        self.index_path = (index_path or settings.FAISS_INDEX_PATH) / "index.bin"
        self.meta_path = meta_path or settings.FAISS_METADATA_PATH

        self.index_path.parent.mkdir(parents=True, exist_ok=True)
        self.meta_path.parent.mkdir(parents=True, exist_ok=True)

        self.index: faiss.Index | None = None
        self._meta: dict[int, dict] = {}   # faiss_id → metadata
        self._next_id: int = 0

        # Load existing index if present
        if self.index_path.exists() and self.meta_path.exists():
            self._load()

    # ── Build / load ──────────────────────────────────────────────────

    def _build_index(self, vectors: np.ndarray) -> faiss.Index:
        """Create IVFFlat index. Falls back to Flat for small corpora."""
        n = vectors.shape[0]
        nlist = min(settings.FAISS_NLIST, max(1, n // 10))

        if n < 1000:
            logger.info(f"Small corpus ({n} vecs) — using Flat index.")
            index = faiss.IndexFlatIP(self.embed_dim)   # inner product = cosine (normalised vecs)
        else:
            logger.info(f"Building IVFFlat index: nlist={nlist}, n={n}")
            quantiser = faiss.IndexFlatIP(self.embed_dim)
            index = faiss.IndexIVFFlat(quantiser, self.embed_dim, nlist, faiss.METRIC_INNER_PRODUCT)
            index.train(vectors)
            index.nprobe = settings.FAISS_NPROBE

        return index

    def _load(self) -> None:
        logger.info(f"Loading FAISS index from {self.index_path}")
        self.index = faiss.read_index(str(self.index_path))

        logger.info(f"Loading metadata from {self.meta_path}")
        with open(self.meta_path) as f:
            for line in f:
                if line.strip():
                    rec = json.loads(line)
                    self._meta[rec["faiss_id"]] = rec
        self._next_id = max(self._meta.keys(), default=-1) + 1
        logger.info(f"Loaded index with {self.index.ntotal} vectors, {len(self._meta)} metadata records.")

    def save(self) -> None:
        faiss.write_index(self.index, str(self.index_path))
        with open(self.meta_path, "w") as f:
            for rec in self._meta.values():
                f.write(json.dumps(rec) + "\n")
        logger.info(f"Saved index ({self.index.ntotal} vectors) → {self.index_path}")

    # ── Add vectors ───────────────────────────────────────────────────

    def add(self, vectors: np.ndarray, posts: list[LinkedInPost]) -> int:
        """
        Add vectors + posts to the index. Builds index if first call.
        Returns number of vectors added.
        """
        assert len(vectors) == len(posts), "vectors and posts must be same length"
        if len(vectors) == 0:
            return 0

        # Assign FAISS IDs
        faiss_ids = list(range(self._next_id, self._next_id + len(vectors)))
        self._next_id += len(vectors)

        # Build or train index if needed
        if self.index is None:
            self.index = self._build_index(vectors)
        elif hasattr(self.index, "is_trained") and not self.index.is_trained:
            self.index.train(vectors)

        # Wrap with IDMap to use custom IDs
        id_index = faiss.IndexIDMap(self.index) if not isinstance(self.index, faiss.IndexIDMap) else self.index
        id_array = np.array(faiss_ids, dtype=np.int64)
        id_index.add_with_ids(vectors, id_array)
        self.index = id_index

        # Store metadata
        for fid, post in zip(faiss_ids, posts):
            self._meta[fid] = _post_to_meta(post, fid)

        logger.info(f"Added {len(vectors)} vectors. Index total: {self.index.ntotal}")
        return len(vectors)

    # ── Search ────────────────────────────────────────────────────────

    def search(
        self,
        query_vector: np.ndarray,
        k: int = 10,
        filters: dict[str, Any] | None = None,
    ) -> list[dict]:
        """
        Search index, return top-k metadata records.
        Optional filters: {"is_hiring": True, "skills": ["python"]}
        """
        if self.index is None or self.index.ntotal == 0:
            logger.warning("Index is empty.")
            return []

        # Fetch 3× k to allow post-filter
        fetch_k = min(k * 3 if filters else k, self.index.ntotal)
        scores, ids = self.index.search(query_vector.astype(np.float32), fetch_k)

        results = []
        for score, fid in zip(scores[0], ids[0]):
            if fid == -1:
                continue
            meta = self._meta.get(int(fid))
            if meta is None:
                continue

            # Apply optional filters
            if filters:
                if "is_hiring" in filters and meta.get("is_hiring") != filters["is_hiring"]:
                    continue
                if "skills" in filters:
                    post_skills = set(meta.get("skills", []))
                    if not any(s in post_skills for s in filters["skills"]):
                        continue
                if "roles" in filters:
                    post_roles = set(meta.get("roles", []))
                    if not any(r in post_roles for r in filters["roles"]):
                        continue
                if "sources" in filters:
                    if meta.get("source") not in filters["sources"]:
                        continue

            results.append({**meta, "score": float(score)})
            if len(results) >= k:
                break

        return results

    # ── TTL eviction ──────────────────────────────────────────────────

    def evict_stale(self, ttl_days: int | None = None) -> int:
        """
        Remove vectors indexed more than `ttl_days` ago.
        Rebuilds index from surviving vectors.
        Returns number of vectors evicted.
        """
        ttl_days = ttl_days or settings.VECTOR_TTL_DAYS
        cutoff = datetime.now(timezone.utc) - timedelta(days=ttl_days)

        stale_ids = []
        for fid, meta in self._meta.items():
            indexed_at = datetime.fromisoformat(meta["indexed_at"])
            if indexed_at < cutoff:
                stale_ids.append(fid)

        if not stale_ids:
            logger.info("No stale vectors to evict.")
            return 0

        logger.info(f"Evicting {len(stale_ids)} stale vectors (TTL={ttl_days}d)")

        # Remove from metadata
        for fid in stale_ids:
            del self._meta[fid]

        # Remove from FAISS index
        stale_array = np.array(stale_ids, dtype=np.int64)
        if hasattr(self.index, "remove_ids"):
            self.index.remove_ids(stale_array)
        else:
            # Fallback: full rebuild from remaining vectors
            self._rebuild_from_meta()

        logger.info(f"Eviction complete. Remaining: {self.index.ntotal} vectors.")
        return len(stale_ids)

    def _rebuild_from_meta(self) -> None:
        """Reconstruct index from metadata (used when FAISS doesn't support remove_ids)."""
        logger.info("Rebuilding index from metadata texts (re-embedding not needed — reconstruct).")
        # NOTE: In production, store raw vectors to GCS for lossless rebuild.
        # For local dev, we reset and mark index for re-ingestion.
        self.index = None
        logger.warning("Index reset — re-ingest required after TTL rebuild.")

    # ── Stats ─────────────────────────────────────────────────────────

    def stats(self) -> dict:
        return {
            "total_vectors": self.index.ntotal if self.index else 0,
            "total_metadata": len(self._meta),
            "embed_dim": self.embed_dim,
            "index_path": str(self.index_path),
            "meta_path": str(self.meta_path),
        }
