linkedin_search/requirements.txt
# ── Core ML / embeddings ──────────────────────────────────────────────────────
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0              # use faiss-gpu if you have a GPU
numpy>=1.26.0
torch>=2.2.0                  # CPU-only: pip install torch --index-url https://download.pytorch.org/whl/cpu

# ── LLM / Groq ───────────────────────────────────────────────────────────────
groq>=0.9.0

# ── LangChain ─────────────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.6
langchain-core>=0.2.0

# ── Scraping ─────────────────────────────────────────────────────────────────
playwright>=1.44.0

# ── FastAPI (for search service, next phase) ──────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# ── Airflow (install separately; heavy dependency) ────────────────────────────
# apache-airflow>=2.9.0
# apache-airflow-providers-google>=10.0.0   # for GCS

# ── GCS (optional) ───────────────────────────────────────────────────────────
# google-cloud-storage>=2.16.0

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0                 # for FastAPI test client
